+++ src/cerbos/sdk/_async/_hub/_store.py
@@ -45,9 +45,28 @@ _MAX_ZIP_SIZE = 15 * 1024 * 1024
 _MIN_ZIP_SIZE = 22
 
 
+def _exclude_errors(thrown_type, thrown_value):
+    if isinstance(thrown_value, grpc.RpcError):
+        status = rpc_status.from_call(thrown_value)
+        if status is None:
+            return False
+
+        return status.code in {
+            code_pb2.ABORTED,
+            code_pb2.CANCELLED,
+            code_pb2.DEADLINE_EXCEEDED,
+        }
+
+    return False
+
+
 class CircuitBreaker(circuitbreaker.CircuitBreaker):
     FAILURE_THRESHOLD = 10
     RECOVERY_TIMEOUT = 60
+    EXPECTED_EXCEPTION = _exclude_errors
+
+
+_CIRCUIT_BREAKER = CircuitBreaker().decorate
 
 
 def handle_store_errors(method):
@@ -146,7 +165,7 @@ class AsyncCerbosHubStoreClient(_AsyncCerbosHubClientBase):
         self._store_stub = store_pb2_grpc.CerbosStoreServiceStub(self._channel)
 
     @handle_store_errors
-    @circuitbreaker.circuit(cls=CircuitBreaker)
+    @_CIRCUIT_BREAKER
     async def replace_files(
         self,
         store_id: str,
@@ -222,7 +241,7 @@ class AsyncCerbosHubStoreClient(_AsyncCerbosHubClientBase):
         return ReplaceFilesResponse(resp)
 
     @handle_store_errors
-    @circuitbreaker.circuit(cls=CircuitBreaker)
+    @_CIRCUIT_BREAKER
     async def replace_files_lenient(
         self,
         store_id: str,
@@ -247,7 +266,7 @@ class AsyncCerbosHubStoreClient(_AsyncCerbosHubClientBase):
             )
 
     @handle_store_errors
-    @circuitbreaker.circuit(cls=CircuitBreaker)
+    @_CIRCUIT_BREAKER
     async def modify_files(
         self,
         store_id: str,
@@ -315,7 +334,7 @@ class AsyncCerbosHubStoreClient(_AsyncCerbosHubClientBase):
         return ModifyFilesResponse(resp)
 
     @handle_store_errors
-    @circuitbreaker.circuit(cls=CircuitBreaker)
+    @_CIRCUIT_BREAKER
     async def modify_files_lenient(
         self,
         store_id: str,
@@ -338,7 +357,7 @@ class AsyncCerbosHubStoreClient(_AsyncCerbosHubClientBase):
             )
 
     @handle_store_errors
-    @circuitbreaker.circuit(cls=CircuitBreaker)
+    @_CIRCUIT_BREAKER
     async def get_files(
         self, store_id: str, file_paths: Iterable[str]
     ) -> GetFilesResponse:
